Fix fastboot device detection under edge cases
The fastboot cli utility returns a status of "fastboot" if the usb interface name is null.
It can only be non-null on Linux, and this theoretically should only happen with a phone running fastbootd, in which case it will be "fastbootd" instead.
The issue occurs due to a bug where reuse of the same struct and failure to reset the string causes the interface name for a different device on the same USB bus to be returned instead.
Since only devices in fastboot mode should be listed anyway, the solution is to not care about the status string and instead just look for the serial number.
Also, fastboot doesn't respect "-s" under the "devices" sub-command.